Move constant-time functions to Subtle class and add necessary MethodImplOptions flags
