+++ dotnet/Miscreant/Utils.cs
@@ -29,8 +29,8 @@ namespace Miscreant
 				input[i] = (byte)((input[i] << 1) | (input[i + 1] >> 7));
 			}
 
-			byte last = (byte)((input[Constants.BlockSize - 1] << 1) ^ ((0 - carry) & Constants.R));
-			input[Constants.BlockSize - 1] = last;
+			byte last = input[Constants.BlockSize - 1];
+			input[Constants.BlockSize - 1] = (byte)((last << 1) ^ Subtle.ConstantTimeSelect(carry, Constants.R, 0));
 		}
 
 		public static void Xor(byte[] source, byte[] destination, int length)
@@ -54,27 +54,6 @@ namespace Miscreant
 			{
 				buffer[i] = 0;
 			}
-		}
-
-		public static bool ConstantTimeEquals(byte[] x, byte[] y, int count)
-		{
-			byte result = 0;
-
-			for (int i = 0; i < count; ++i)
-			{
-				result |= (byte)(x[i] ^ y[i]);
-			}
-
-			return result == 0;
-		}
-
-		/// <summary>
-		/// ConstantTimeSelect returns x if v is 1 and y if v is 0.
-		/// See <see href="https://golang.org/src/crypto/subtle/constant_time.go">constant_time.go</see> for more details.
-		/// </summary>
-		public static int ConstantTimeSelect(int v, int x, int y)
-		{
-			return ~(v - 1) & x | (v - 1) & y;
 		}
 
 		public static int Ceil(int dividend, int divisor)
